fix(PDFEditor): fix OpenJPEG image decoding error
Opening a PDF with JPEG 2000 images causes these warnings:

  Warning: JpxImage#instantiateWasm: UnknownErrorException: Ensure that the `wasmUrl` API parameter is provided.
  Warning: JpxImage#getJsModule: TypeError: Failed to resolve module specifier 'nullopenjpeg_nowasm_fallback.js'
  Warning: Unable to decode image "img_p0_2": "JpxError: OpenJPEG failed to initialize".

In version 5.3.31 [1], the OpenJPEG wasm fallback logic was removed in
PR mozilla/pdf.js#19923. The wasmUrl option is required since v5 [2] to
decode JPEG 2000 images.

There are two solutions to this:

	1. Update Webpack config to copy the
	   node_modules/pdfjs-dist/wasm folder to static folder.
	2. Load the openjpeg.wasm file from a CDN like jsDelivr.

Method 2 is preferred because jsDelivr CDN sets the Cache-Control header
and these resources will be cached by the browser correctly.

This commit fixes the bug by providing the wasmUrl to jsDelivr CDN. We also
provide the cMapUrl and iccUrl to avoid similar warnings from happening
in the future.

[1]: https://github.com/mozilla/pdf.js/releases/tag/v5.3.31
[2]: https://github.com/mozilla/pdf.js/releases/tag/v5.0.375

Signed-off-by: Eng Zer Jun <engzerjun@gmail.com>